src/django/requirements.txt
black==22.3.0
boto3==1.24.3
django==3.2.16
django-ecsmanage==2.0.1
django-extensions==3.1.5
django-simple-history==3.1.1
django-spa==0.3.6
django-storages==1.12.3
django-watchman==1.3.0
djangorestframework==3.13.1
djangorestframework-gis==1.0
factory-boy==3.2.1
filetype==1.0.13
gunicorn==20.1.0
psycopg2==2.9.3
python-jose==3.3
pytz==2022.1
requests==2.27.1
